edited cli.py to be compatible with windows
